fix: 🚑️ modified auth callaback (#166)
caused continuous invoking of `GET /api/auth/session`

Related: #166
