+++ src/app/layout.tsx
@@ -1,8 +1,7 @@
+import AuthProvider from '@/components/AuthProvider';
 import type { Metadata } from 'next';
 import { Josefin_Sans } from 'next/font/google';
 import './globals.css';
-
-import AuthProvider from '@/components/AuthProvider';
 
 const josefinSans = Josefin_Sans({
   subsets: ['latin'],
